Doing proper mock cleanup
Now it checks if there are any mounted dirs inside the chroot and tries to
umount before removing. It will fail if unable to umount.

Also the cleanup has been put as post-build task, so it should always run.
Changed the logs tarball extension to more standard tgz

+++ jobs/confs/shell-scripts/mock_cleanup.sh
@@ -1,12 +1,70 @@
 #!/bin/bash -xe
 echo "shell-scripts/mock_cleanup.sh"
-# remove chroot to free space
+
+shopt -s nullglob
+
+# Archive the logs, we want them anyway
+pushd "$WORKSPACE"/exported-artifacts
+logs=(*log *_pkgs "$WORKSPACE"/*log)
+if [[ "$logs" ]]; then
+    tar cvzf logs.tgz "${logs[@]}"
+    rm -f *log *_pkgs
+fi
+popd
+
+failed=false
+
+mock_dir="$WORKSPACE/mock"
+# stop any processes running inside the chroot
+chroots=("$WORKSPACE"/mock/*)
+for chroot in "${chroots[@]}"; do
+    echo "Cleaning up chroot $chroot"
+    ### Generate the mock configuration
+    pushd "$WORKSPACE"/jenkins/mock_configs
+    mock_conf="${chroot##*/}"
+    base_conf="${mock_conf%%x86_64*}x86_64"
+    echo "#### Generating mock configuration"
+    ./mock_genconfig \
+        --name="$mock_conf" \
+        --base="$base_conf.cfg" \
+        --option="basedir=$WORKSPACE/mock/" \
+    > "$mock_conf.cfg"
+    cat "$mock_conf.cfg"
+    popd
+
+    my_mock="/usr/bin/mock"
+    my_mock+=" --configdir=$WORKSPACE/jenkins/mock_configs"
+    my_mock+=" --root=$mock_conf"
+
+    #TODO: investigate why mock --clean fails to umount certain dirs sometimes,
+    #so we can use it instead of manually doing all this.
+    echo "Killing all mock orphan processes, if any."
+    $my_mock \
+        --orphanskill \
+    || {
+        echo "ERROR:  Failed to kill orphans on $chroot."
+        failed=true
+    }
+
+    mounts=($(mount | awk '{print $3}' | grep "$chroot")) || :
+    if [[ "$mounts" ]]; then
+        echo "Found mounted dirs inside the chroot $chroot. Trying to umount."
+    fi
+    for mount in "${mounts[@]}"; do
+        sudo umount "$mount" \
+        || {
+            echo "ERROR:  Failed to umount $mount."
+            failed=true
+        }
+    done
+done
+
+if $failed; then
+    echo "Aborting."
+    exit 1
+fi
 sudo rm -Rf mock mock-cache
+
 # remove mock system cache, we will setup proxies to do the caching and this
 # takes lots of space between runs
 sudo rm -Rf /var/cache/mock/*
-# compress logs
-pushd "$WORKSPACE"/exported-artifacts
-shopt -s nullglob
-tar cvjf logs.tbz *log *_pkgs "$WORKSPACE"/*log
-rm -f *log *_pkgs
